requirements.txt
apispec==6.6.1
apispec-webframeworks==1.1.0
attrs==23.2.0
blinker==1.8.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.7
exceptiongroup==1.2.1
flasgger==0.9.7.1
Flask==3.0.3
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
greenlet==3.0.3
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
marshmallow==3.21.2
marshmallow-sqlalchemy==1.0.0
mistune==3.0.2
packaging==24.0
pluggy==1.5.0
pycparser==2.22
PyMySQL==1.1.1
pytest==8.2.1
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
six==1.16.0
SQLAlchemy==2.0.30
tomli==2.0.1
typing_extensions==4.11.0
Werkzeug==3.0.3
